agent_service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
strands-agents
strands-agents-tools
google-genai
ddgs
requests
beautifulsoup4
uv